added security

+++ src/main/java/com/dotoku/carhop/service/RequestService.java
@@ -6,15 +6,18 @@ import com.dotoku.carhop.entity.HopSession;
 import com.dotoku.carhop.entity.HopUser;
 import com.dotoku.carhop.entity.RequestStatus;
 import com.dotoku.carhop.repository.RequestRepository;
-import com.dotoku.carhop.repository.ReviewRepository;
 import com.dotoku.carhop.repository.SessionRepository;
 import com.dotoku.carhop.repository.UserRepository;
+import com.dotoku.carhop.security.SecurityUtil;
 import jakarta.persistence.EntityNotFoundException;
+import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDateTime;
+import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
@@ -55,6 +58,11 @@ public class RequestService {
         HopRequest hopRequest = requestRepository.findById(requestId)
                 .orElseThrow(() -> new EntityNotFoundException("Request Not found!"));
 
+        String loggedInEmail = SecurityUtil.getLoggedInEmail();
+        if (!hopRequest.getHopUser().getEmail().equals(loggedInEmail)) {
+            throw new IllegalStateException("You are not authorized to delete this request.");
+        }
+
         requestRepository.delete(hopRequest);
 
         return ResponseEntity.ok("Request Canceled!");
@@ -64,6 +72,11 @@ public class RequestService {
 
         HopRequest hopRequest = requestRepository.findById(requestId)
                 .orElseThrow(() -> new EntityNotFoundException("Request Not found!"));
+
+        String loggedInEmail = SecurityUtil.getLoggedInEmail();
+        if (!hopRequest.getHopSession().getHopUser().getEmail().equals(loggedInEmail)) {
+            throw new IllegalStateException("You are not authorized to approve this session.");
+        }
 
         hopRequest.setStatus(RequestStatus.APPROVED.toString());
 
@@ -77,9 +90,60 @@ public class RequestService {
         HopRequest hopRequest = requestRepository.findById(requestId)
                 .orElseThrow(() -> new EntityNotFoundException("Request Not found!"));
 
+        String loggedInEmail = SecurityUtil.getLoggedInEmail();
+        if (!hopRequest.getHopSession().getHopUser().getEmail().equals(loggedInEmail)) {
+            throw new IllegalStateException("You are not authorized to reject this request.");
+        }
+
         hopRequest.setStatus(RequestStatus.REJECTED.toString());
         requestRepository.delete(hopRequest);
 
         return ResponseEntity.ok("Request Rejected!");
     }
+
+    public ResponseEntity<List<HopRequestDto>> getRequestsByUserId(Long userId) {
+        String loggedInEmail = SecurityUtil.getLoggedInEmail();
+
+        HopUser hopUser = userRepository.findById(userId)
+                .orElseThrow(() -> new EntityNotFoundException("User not found!"));
+
+        if (!hopUser.getEmail().equals(loggedInEmail)) {
+            throw new IllegalStateException("You are not authorized to view these requests.");
+        }
+
+        List<HopRequest> requests = requestRepository.findByUserId(userId);
+
+        List<HopRequestDto> hopRequestDtos = requests.stream()
+                .map(this::mapRequestEntityToDto)
+                .collect(Collectors.toList());
+
+        return ResponseEntity.ok(hopRequestDtos);
+    }
+
+
+    public ResponseEntity<HopRequestDto> getRequestByRequestId(Long requestId) {
+        String loggedInEmail = SecurityUtil.getLoggedInEmail();
+
+        HopRequest hopRequest = requestRepository.findById(requestId)
+                .orElseThrow(() -> new EntityNotFoundException("Request not found!"));
+
+        if (!hopRequest.getHopUser().getEmail().equals(loggedInEmail)) {
+            throw new IllegalStateException("You are not authorized to view these requests.");
+        }
+
+        HopRequestDto hopRequestDto = mapRequestEntityToDto(hopRequest);
+
+        return ResponseEntity.ok(hopRequestDto);
+    }
+
+    public HopRequestDto mapRequestEntityToDto(HopRequest hopRequest){
+        HopRequestDto hopRequestDto = new HopRequestDto();
+        hopRequestDto.setHopSessionId(hopRequest.getHopSession().getId());
+        hopRequestDto.setHopUserId(hopRequest.getHopUser().getId());
+        hopRequestDto.setRequestTime(hopRequest.getRequestTime());
+        hopRequestDto.setStatus(hopRequest.getStatus());
+        hopRequestDto.setId(hopRequest.getId());
+        return hopRequestDto;
+    }
+
 }
